re PR fortran/32770 ([Meta-bug] -fdefault-integer-8 issues)
2007-08-02  Thomas Koenig  <tkoenig@gcc.gnu.org>

	PR fortran/32770
	* gfortran.dg/unf_read_corrupted_1.f90:  Make all kinds
	explicit so test passes with -fdefault-integer-8.
	* gfortran.dg/unformatted_recl_1.f90:  Likewise.
	* gfortran.dg/unformatted_subrecord_1.f90:  Likewise.

From-SVN: r127168

+++ gcc/testsuite/gfortran.dg/unformatted_recl_1.f90
@@ -1,7 +1,7 @@
 ! { dg-do run } 
 ! PR31099 Runtime error on legal code using RECL
 program test
-  integer :: a, b
+  integer(kind=4) :: a, b
   a=1
   b=2
   open(10, status="scratch", form="unformatted", recl=8)
